\name{fdr.table}

\alias{fdr.table}

\title{TANOVA False Discovery Table}

\description{It's an internal function. The function compare ANOVA F-statistics to null distribution generated by Bootstrap resampling to estimate False Discovery Rate}

\usage{fdr.table(obj)}

\arguments{
\item{obj}{List object containing at least two components: one is the F-statistics for ANOVA, the other is the bootstrap generated null distribution of F-statistics}
}

\value{
\item{table}{a matrix contains estimated false discovery rate at various cutoffs}
}

\author{Baiyu Zhou \email{zhouby98@stanford.edu} & Weihong xu \email{weihongx@stanford.edu}}

\seealso{
  \code{\link{tanova}}
}

\keyword{TANOVA}
