
\name{group.ix}

\alias{group.ix}

\title{This is an internal function}

\description{
  This is an internal function.}

\usage{
  group.ix(f1,f2)
}

\arguments{
\item{f1}{a vector of length equal to the number of arrays. Each entry indicates the level of the first factor for corresponding array. The values of f1 should be 1,2,3,\dots}

\item{f2}{a vector of length equal to the number of arrays. Each entry indicates the 
level of the second factor for the corresponding array. The values of f2 should be 1,2,3,\dots
If the experimental has only one factor, let f2=0.}
}

%\value{}

\author{Baiyu Zhou \email{zhouby98@stanford.edu} & Weihong xu \email{weihongx@stanford.edu}}

\keyword{TANOVA}
