%####################################################################
% TANOVA:% trigammainverse.Rd
% Function: Write Documentation
% Weihong Xu & BaiyuZhou
% Aug 9, 2009
%####################################################################

\name{trigammaInverse}

\alias{trigammaInverse}

\title{Trigamma Inverse Function}

\description{This is an internal function.}

\usage{trigammaInverse(x)}

\arguments{
\item{x}{internal object}
}


\author{Baiyu Zhou \email{zhouby98@stanford.edu} & Weihong xu \email{weihongx@stanford.edu}}

\seealso{
  \code{\link{tanova}}
}

\keyword{TANOVA}
